Autogen/code/Autogen_GraphRAG_Ollama/requirements.txt
litellm[proxy]
ollama
pyautogen[retrievechat]
tiktoken
chainlit
graphrag
marker-pdf
torch
